requirements.txt
# Core ML/DL Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
efficientnet_pytorch>=0.7.1

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Handling
pandas>=2.0.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking and Logging
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration Management
PyYAML>=6.0
omegaconf>=2.3.0

# Progress Bars and CLI
tqdm>=4.65.0
click>=8.1.0

# Image Processing and Fonts
fonttools>=4.40.0
freetype-py>=2.4.0

# Text Processing (Unicode normalization)
unicodedata2>=15.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipython>=8.14.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Utilities
scikit-learn>=1.3.0
scikit-image>=0.21.0 